LPS-106487 Move Resize logic from Row to RowWithControls
